Updates treesitter query to address 'invalid node type' error
the current (July 2025) version of nvim-treesitter was raising errors on multiple
lines of the query in this module.  It appears that 'function' is no longer a valid
node type; using the [tree-sitter playgroud](https://tree-sitter.github.io/tree-sitter/7-playground.html)
I determined that 'function_expression' is now used for this node.